refactor: TuripStreamResult 및 StreamFailureAction을 data 패키지로 이동 및 분리
