+++ android/app/src/main/java/com/on/turip/data/result/StreamFailureAction.kt
@@ -1,0 +1,8 @@
+package com.on.turip.data.result
+
+enum class StreamFailureAction {
+    Retry,
+    Stop,
+    FatalTokenExpired,
+    FatalForbidden,
+}
